Moved factories to separate folders
