+++ src/game/graphics_handler/entities/factories/entity_factories.js
@@ -1,9 +1,9 @@
 import * as PIXI from "https://cdnjs.cloudflare.com/ajax/libs/pixi.js/8.1.5/pixi.mjs";
-import { GridEntity } from "./grid_entity.js";
-import { GraphicsEntity } from "./graphics_entity.js";
-import { BackgroundEntity } from "./background_entity.js";
-import { GridObjectEntity } from "./grid_object_entity.js";
-import { TextBoxEntity } from "./textbox_entity.js";
+import { GridEntity } from "../grid_entity.js";
+import { GraphicsEntity } from "../graphics_entity.js";
+import { BackgroundEntity } from "../background_entity.js";
+import { GridObjectEntity } from "../grid_object_entity.js";
+import { TextBoxEntity } from "../textbox_entity.js";
 
 /**
  * @private
