+++ src/game/graphics_handler/entity_skins/factories/entity_skin_factories.js
@@ -1,4 +1,4 @@
-import { EntitySkin } from "./entity_skin.js";
+import { EntitySkin } from "../entity_skin.js";
 
 export function createCharacterBunnySkin(skinName, theme, assets) {
     let skinOptions = {
